Add files via upload
Since the time when pandemic started, department’s sales appear to be unpredictable. Thus, to better understand the direction of department performance, it makes sense first to explore the sales during the pre-pandemic normal times and find the prior associations.
So the purpose of this analysis is to predict sales based on the time series using historical data from 2016 to pre-pandemic period (February, 2020). Further compare predicted values with the actual data and make the determination how applicable Winter-Holts model can be.